R-U-N-master/py_belt_exam_20Dec16/requirements.txt
bcrypt==3.1.1
cffi==1.9.1
click==6.6
colorama==0.3.7
Django==1.10.4
django-bootstrap3==7.1.0
django-debug-toolbar==1.6
django-floppyforms==1.7.0
docopt==0.6.2
ez-setup==0.9
feedparser==5.2.1
gevent==1.1.2
gitsome==0.6.0
greenlet==0.4.11
gunicorn==19.6.0
Jinja2==2.8
MarkupSafe==0.23
numpydoc==0.6.0
ply==3.9
prompt-toolkit==1.0.9
pycparser==2.17
Pygments==2.1.3
pytz==2016.7
requests==2.12.3
six==1.10.0
SQLAlchemy==1.1.4
sqlparse==0.2.2
uritemplate==3.0.0
uritemplate.py==3.0.2
wcwidth==0.1.7
Werkzeug==0.11.11
